The code contains 2 parts. VUI and Backend.
The lambda codes are in Planet_pluto lambda.js which contains the backend, the code is designed in such a that it follows an incremental model for the further addition of game features and update in any type of conversation design. I would encourage the user to give an addition to the code and even collaborate with me for further skill development. This will help us to build better, faster and error-free skills.

The Planet_pluto.jason file contains the Voice User Interface part which needs to be uploaded to Amazon developers skill interaction section with the help of which the users can interact with the Alexa skill. This is simple machine generated code which will intern generate a Voice interaction model that can be updated easily.